+++ core/src/com/unciv/logic/civilization/managers/TechManager.kt
@@ -525,12 +525,12 @@ class TechManager : IsPartOfGameInfoSerialization {
     }
 
     fun getBestRoadAvailable(): RoadStatus {
-        val railroadImprovement = RoadStatus.Railroad.improvement(getRuleset())  // May not exist in mods
-        if (railroadImprovement != null && (railroadImprovement.techRequired==null || isResearched(railroadImprovement.techRequired!!)))
+        val railroadImprovement = getRuleset().railroadImprovement  // May not exist in mods
+        if (railroadImprovement != null && (railroadImprovement.techRequired == null || isResearched(railroadImprovement.techRequired!!)))
             return RoadStatus.Railroad
 
-        val roadImprovement = RoadStatus.Road.improvement(getRuleset())
-        if (roadImprovement != null && (roadImprovement.techRequired==null || isResearched(roadImprovement.techRequired!!)))
+        val roadImprovement = getRuleset().roadImprovement
+        if (roadImprovement != null && (roadImprovement.techRequired == null || isResearched(roadImprovement.techRequired!!)))
             return RoadStatus.Road
 
         return RoadStatus.None