+++ core/src/com/unciv/models/Spy.kt
@@ -8,6 +8,9 @@ import com.unciv.logic.civilization.Civilization
 import com.unciv.logic.civilization.NotificationCategory
 import com.unciv.logic.civilization.NotificationIcon
 import com.unciv.logic.civilization.managers.EspionageManager
+import com.unciv.models.ruleset.unique.StateForConditionals
+import com.unciv.models.ruleset.unique.Unique
+import com.unciv.models.ruleset.unique.UniqueType
 import kotlin.random.Random
 
 
@@ -100,8 +103,10 @@ class Spy() : IsPartOfGameInfoSerialization {
                     return
                 }
                 val techStealCost = stealableTechs.maxOfOrNull { civInfo.gameInfo.ruleset.technologies[it]!!.cost }!!
+                var progressThisTurn = getLocation()!!.cityStats.currentCityStats.science
                 // 33% spy bonus for each level
-                val progressThisTurn = getLocation()!!.cityStats.currentCityStats.science * (rank + 2f) / 3f
+                progressThisTurn *= (rank + 2f) / 3f
+                progressThisTurn *= getEfficiencyModifier().toFloat()
                 progressTowardsStealingTech += progressThisTurn.toInt()
                 if (progressTowardsStealingTech > techStealCost) {
                     stealTech()
@@ -120,7 +125,7 @@ class Spy() : IsPartOfGameInfoSerialization {
                 val oldSpyName = name
                 name = espionageManager.getSpyName()
                 action = SpyAction.None
-                civInfo.addNotification("We have recruited a new spy name [$name] after [$oldSpyName] was killed.", 
+                civInfo.addNotification("We have recruited a new spy name [$name] after [$oldSpyName] was killed.",
                     NotificationCategory.Espionage, NotificationIcon.Spy)
             }
             SpyAction.CounterIntelligence -> {
@@ -129,7 +134,7 @@ class Spy() : IsPartOfGameInfoSerialization {
                 // Once turnRemainingForAction is <= 0 the spy won't be considered to be doing work any more
                 --turnsRemainingForAction
                 return
-            } 
+            }
             else -> return // Not implemented yet, so don't do anything
         }
     }
@@ -157,7 +162,6 @@ class Spy() : IsPartOfGameInfoSerialization {
         // Subtract the experience of the counter inteligence spies
         val defendingSpy = city.civ.espionageManager.getSpyAssignedToCity(city)
         spyResult += defendingSpy?.getSkillModifier() ?: 0
-        //TODO: Add policies modifier here
 
         val detectionString = when {
             spyResult < 0 -> null // Not detected
@@ -228,7 +232,7 @@ class Spy() : IsPartOfGameInfoSerialization {
         action = SpyAction.Moving
         turnsRemainingForAction = 1
     }
-    
+
     fun canMoveTo(city: City): Boolean {
         if (getLocation() == city) return true
         if (!city.getCenterTile().isVisible(civInfo)) return false
@@ -238,7 +242,7 @@ class Spy() : IsPartOfGameInfoSerialization {
     fun isSetUp() = action !in listOf(SpyAction.Moving, SpyAction.None, SpyAction.EstablishNetwork)
 
     // Only returns true if the spy is doing a helpful and implemented action
-    fun isDoingWork(): Boolean { 
+    fun isDoingWork(): Boolean {
         if (action == SpyAction.StealingTech || action == SpyAction.EstablishNetwork || action == SpyAction.Moving) return true
         if (action == SpyAction.RiggingElections && !civInfo.isAtWarWith(getLocation()!!.civ)) return true
         if (action == SpyAction.CounterIntelligence && turnsRemainingForAction > 0) return true
@@ -260,9 +264,9 @@ class Spy() : IsPartOfGameInfoSerialization {
 
     fun levelUpSpy() {
         //TODO: Make the spy level cap dependent on some unique
-        if (rank >= 3) return 
+        if (rank >= 3) return
         if (getLocation() != null) {
-            civInfo.addNotification("Your spy [$name] has leveled up!", getLocation()!!.location, 
+            civInfo.addNotification("Your spy [$name] has leveled up!", getLocation()!!.location,
                 NotificationCategory.Espionage, NotificationIcon.Spy)
         } else {
             civInfo.addNotification("Your spy [$name] has leveled up!",
@@ -275,6 +279,32 @@ class Spy() : IsPartOfGameInfoSerialization {
         return getSpyRank() * 30
     }
 
+    /**
+     * Gets a friendly and enemy efficiency uniques for the spy at the location
+     * @return a value centered around 100 for the work efficiency of the spy, won't be negative
+     */
+    fun getEfficiencyModifier(): Double {
+        lateinit var friendlyUniques: Sequence<Unique>
+        lateinit var enemyUniques: Sequence<Unique>
+        if (getLocation() != null) {
+            val city = getLocation()!!
+            if (city.civ == civInfo) {
+                friendlyUniques = city.getMatchingUniques(UniqueType.SpyEffectiveness, StateForConditionals(city), includeCivUniques = true)
+                enemyUniques = sequenceOf()
+            } else {
+                friendlyUniques = civInfo.getMatchingUniques(UniqueType.SpyEffectiveness)
+                enemyUniques = city.getMatchingUniques(UniqueType.EnemySpyEffectiveness, StateForConditionals(city), includeCivUniques = true)
+            }
+        } else {
+            friendlyUniques = civInfo.getMatchingUniques(UniqueType.SpyEffectiveness)
+            enemyUniques = sequenceOf()
+        }
+        var totalEfficiency = 1.0
+        totalEfficiency *= (100.0 + friendlyUniques.sumOf { it.params[0].toInt() }) / 100
+        totalEfficiency *= (100.0 + enemyUniques.sumOf { it.params[0].toInt() }) / 100
+        return totalEfficiency.coerceAtLeast(0.0)
+    }
+
     fun killSpy() {
         // We don't actually remove this spy object, we set them as dead and let them revive
         moveTo(null)
@@ -282,6 +312,6 @@ class Spy() : IsPartOfGameInfoSerialization {
         turnsRemainingForAction = 5
         rank = 1
     }
-    
+
     fun isAlive(): Boolean = action != SpyAction.Dead
 }